Guess the decimals on the numberformatter
