+++ app/Http/Livewire/PriceTicker.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace App\Http\Livewire;
 
-use App\Facades\Network;
-use App\Services\CryptoCompare;
-use App\Services\Settings;
-use Illuminate\View\View;
 use Livewire\Component;
+use App\Facades\Network;
+use Illuminate\View\View;
+use App\Services\Settings;
+use App\Services\CryptoCompare;
+use App\Services\NumberFormatter;
 
 final class PriceTicker extends Component
 {
@@ -28,9 +29,16 @@ final class PriceTicker extends Component
 
     public function setValues(): void
     {
-        $this->price = number_format(CryptoCompare::price(Network::currency(), Settings::currency()), 2);
+        $this->price = $this->getPriceFormatted();
         $this->from  = Network::currency();
         $this->to    = Settings::currency();
+    }
+
+    private function getPriceFormatted(): string
+    {
+        $price = CryptoCompare::price(Network::currency(), Settings::currency());
+
+        return NumberFormatter::currency($price, Settings::currency());
     }
 
     public function render(): View
