+++ tests/Unit/Services/ExchangeRateTest.php
@@ -19,10 +19,3 @@ it('should convert with the current rate', function () {
 
     expect(ExchangeRate::now(1))->toBe(10.0);
 });
-
-it('should use two decimals for a fiat currency', function () {
-    expect(ExchangeRate::decimalsFor('USD'))->toBe(2);
-    expect(ExchangeRate::decimalsFor('BTC'))->toBe(8);
-    expect(ExchangeRate::decimalsFor('ETH'))->toBe(8);
-    expect(ExchangeRate::decimalsFor('LTC'))->toBe(8);
-});
