+++ tests/Unit/Services/NumberFormatterTest.php
@@ -58,3 +58,11 @@ it('should format a number with a K notation', function () {
     assertMatchesSnapshot(NumberFormatter::currencyShortNotation(100000000));
     assertMatchesSnapshot(NumberFormatter::currencyShortNotation(104910000));
 });
+
+
+it('should use two decimals for a fiat currency', function () {
+    expect(NumberFormatter::decimalsFor('USD'))->toBe(2);
+    expect(NumberFormatter::decimalsFor('BTC'))->toBe(8);
+    expect(NumberFormatter::decimalsFor('ETH'))->toBe(8);
+    expect(NumberFormatter::decimalsFor('LTC'))->toBe(8);
+});
